Add files via upload
This project is a simple CRUD operations based demo having employee-department theme. This demo uses Angular as client side and Laravel as server side for development. The main feature of this demo is that features provided by the listing page. The listing page has following features:-

1). Pagination
2). Dynamic records per page
3). Column wise sorting
4). Column wise searching
5). Global searching
6). Dynamic columns
7). Draggable columns

Apart from these features, Edit, Add and Delete of employees can also be found in this demo.